+++ src/buffer_cache/large_buf.hpp
@@ -95,7 +95,7 @@ public:
 
     int64_t get_num_segments();
 
-    uint16_t segment_size(int ix);
+    uint16_t segment_size(int64_t ix);
 
     const byte *get_segment(int64_t num, uint16_t *seg_size);
     byte *get_segment_write(int64_t num, uint16_t *seg_size);
